requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
click==7.1.1
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
Flask==1.1.2
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
imageio==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==2.2.2
networkx==2.1
numpy==1.18.3
opencv-python-headless==4.2.0.34
pandas==0.23.4
Pillow==5.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.18
pygit==0.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
PyYAML==5.1
requests==2.23.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.24.0
Werkzeug==1.0.1
